289: Cleaned up and standardized directory structure, and refactored code to match.
