+++ rfInsight/index.js
@@ -2,7 +2,7 @@ const express = require("express");
 const hbs = require("hbs");
 const http = require("http");
 const baseRoutes = require("./routes/baseRoutes");
-var app_controller = require("./Controllers/AppController");
+var app_controller = require("./controllers/AppController");
 
 let router = express.Router();
 /*
