+++ rfInsight/resources/js/dataGenerator/PatternFuncs.js
