+++ rfInsight/resources/js/dataGenerator/config.js
