+++ rfInsight/resources/js/dataGenerator/redfish/v1/TelemetryService/MetricReports/CPUMetrics/index.json
