+++ rfInsight/resources/js/dataGenerator/redfish/v1/TelemetryService/MetricReports/MemoryMetrics/index.json
