+++ rfInsight/resources/js/dataGenerator/redfish/v1/TelemetryService/index.json
