+++ rfInsight/resources/js/dataGenerator/redfish/v1/explorer_config.json
