+++ rfInsight/resources/js/dataGenerator/redfish/v1/index.json
