+++ rfInsight/resources/js/index.js
