+++ rfInsight/resources/js/redfishTools.js
