+++ rfInsight/resources/js/util.js
