+++ rfInsight/resources/panels/graph1.json
