+++ rfInsight/resources/panels/plotLy1.json
