+++ rfInsight/resources/panels/radar1.json
