+++ rfInsight/resources/panels/tempGauge1.json
