+++ rfInsight/views/chart.hbs
