+++ rfInsight/views/dataGeneratorUI.hbs
@@ -2,7 +2,7 @@
 <html>
 	<head>
 		<link rel="stylesheet" href="/css/app.css" />
-   		<script src="../Resources/js/index.js"></script>
+   		<script src="../resources/js/index.js"></script>
 	    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
    		<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.5.0/css/all.css" integrity="sha384-B4dIYHKNBt8Bc12p+WXckhzcICo0wtJAoU8YZTY5qE0Id1GSseTk6S+L3BlXeVIU"
           crossorigin="anonymous" />
