+++ rfInsight/views/partials/footer.hbs
