+++ rfModeller/PatternFuncs.js
