+++ rfModeller/__test__/PatternFuncs.test.js
